New commit related to the previous one.

+++ src/hitables/movingsphere.h
@@ -48,7 +48,7 @@ class movingSphere: public hitable
 
             if (discriminant > 0)
             {
-                float temp = (-b - sqrt(discriminant))/a;
+                float temp = (-b - static_cast<float>(sqrt(static_cast<double>(discriminant))))/a;
                 if (temp < tMax && temp > tMin)
                 {
                     rec.time = temp;
@@ -57,7 +57,7 @@ class movingSphere: public hitable
                     rec.matPtr = matPtr;
                     return true;
                 }
-                temp = (-b + sqrt(discriminant))/a;
+                temp = (-b + static_cast<float>(sqrt(static_cast<double>(discriminant))))/a;
                 if (temp < tMax && temp > tMin)
                 {
                     rec.time = temp;
