+++ src/util/vec3.h
@@ -163,7 +163,7 @@ inline CUDA_HOSTDEV vec3& vec3::operator/=(const float x)
 
 inline CUDA_HOSTDEV float vec3::length() const
 {
-    return sqrt(e[0]*e[0]+e[1]*e[1]+e[2]*e[2]);
+    return static_cast<float>(sqrt(static_cast<double>(e[0]*e[0]+e[1]*e[1]+e[2]*e[2])));
 }
 
 inline CUDA_HOSTDEV float vec3::squaredLength() const
@@ -173,7 +173,7 @@ inline CUDA_HOSTDEV float vec3::squaredLength() const
 
 inline CUDA_HOSTDEV void vec3::makeUnitVector()
 {
-    float k = 1.0f / (sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]));
+    float k = 1.0f / (static_cast<float>(sqrt(static_cast<double>(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]))));
     e[0] *= k;
     e[1] *= k;
     e[2] *= k;
